Add interface in Storage; fixed Storage.getBook, Storage.updateBook,Storage.newBook;
